Update pretreat.m
Had to rewrite some parts of plotting functions bc I made the inputs too confusing. Also moved the normalization of the dataset to the end to make it easier to read it at various steps (still normalized for parts that need it, just not saved into the variable). Added removal of time samples whenever data samples are removed to keep them in sync. Also removed the shift of the output variable back to the sample before it. It makes performance worse, so I assume they must have already adjusted it. Commented out some of the histograms bc I didn't find them useful at this stage.